Time required to load the whole sdk is high. AWS has over 200 gems in the sdk. We only need 40 or so.

➜  ~  time ruby -e "require 'aws-sdk'"
ruby -e "require 'aws-sdk'"  28.13s user 5.96s system 99% cpu 34.165 total
➜  ~  time ruby -e "require 'aws-sdk-s3'"
ruby -e "require 'aws-sdk-s3'"  0.37s user 0.17s system 93% cpu 0.572 total